Unify notification bell: Popover preview + Sheet for full list
Replace two separate notification UIs (global popover + trip-specific sheet)
with a single bell in the app header. Quick preview in a Popover, "View all"
opens a Sheet with infinite scroll. Remove trip notification bell from trip
detail page. Fix bell icon off-center due to aria-live span layout.
