+++ apps/web/src/app/(app)/trips/[id]/trip-detail-content.test.tsx
@@ -223,13 +223,8 @@ vi.mock("@/components/trip/members-list", () => ({
   ),
 }));
 
-// Mock TripNotificationBell and TripSettingsButton components
+// Mock TripSettingsButton component
 vi.mock("@/components/notifications", () => ({
-  TripNotificationBell: ({ tripId }: { tripId: string }) => (
-    <div data-testid="trip-notification-bell" data-trip-id={tripId}>
-      Trip Notification Bell
-    </div>
-  ),
   TripSettingsButton: ({ tripId }: { tripId: string }) => (
     <div data-testid="trip-settings-button" data-trip-id={tripId}>
       Trip Settings
@@ -2020,123 +2015,4 @@ describe("TripDetailContent", () => {
     });
   });
 
-  describe("trip notification bell", () => {
-    it("renders TripNotificationBell with correct tripId for organizer", async () => {
-      mockUseTripDetail.mockReturnValue({
-        data: mockTripDetail,
-        isPending: false,
-        isError: false,
-        error: null,
-        refetch: vi.fn(),
-      });
-
-      render(
-        <Suspense fallback={null}>
-          <TripDetailContent tripId="trip-123" />
-        </Suspense>,
-      );
-
-      await waitFor(() => {
-        const bell = screen.getByTestId("trip-notification-bell");
-        expect(bell).toBeDefined();
-        expect(bell.getAttribute("data-trip-id")).toBe("trip-123");
-      });
-    });
-
-    it("renders TripNotificationBell for non-organizer members", async () => {
-      const regularUser = { ...mockUser, id: "user-789" };
-      mockUseAuth.mockReturnValue({ user: regularUser });
-      mockUseTripDetail.mockReturnValue({
-        data: { ...mockTripDetail, isOrganizer: false },
-        isPending: false,
-        isError: false,
-        error: null,
-        refetch: vi.fn(),
-      });
-
-      render(
-        <Suspense fallback={null}>
-          <TripDetailContent tripId="trip-123" />
-        </Suspense>,
-      );
-
-      await waitFor(() => {
-        const bell = screen.getByTestId("trip-notification-bell");
-        expect(bell).toBeDefined();
-        expect(bell.getAttribute("data-trip-id")).toBe("trip-123");
-      });
-
-      // Non-organizer should not see edit/invite but should see bell
-      expect(screen.queryByText("Edit trip")).toBeNull();
-      expect(screen.queryByText("Invite")).toBeNull();
-    });
-
-    it("does not render TripNotificationBell in preview mode", async () => {
-      const previewTrip = {
-        ...mockTripDetail,
-        isPreview: true,
-        userRsvpStatus: "no_response" as const,
-        isOrganizer: false,
-      };
-      mockUseTripDetail.mockReturnValue({
-        data: previewTrip,
-        isPending: false,
-        isError: false,
-        error: null,
-        refetch: vi.fn(),
-      });
-
-      render(
-        <Suspense fallback={null}>
-          <TripDetailContent tripId="trip-123" />
-        </Suspense>,
-      );
-
-      await waitFor(() => {
-        expect(screen.getByTestId("trip-preview")).toBeDefined();
-      });
-
-      expect(screen.queryByTestId("trip-notification-bell")).toBeNull();
-    });
-
-    it("does not render TripNotificationBell in error state", async () => {
-      mockUseTripDetail.mockReturnValue({
-        data: undefined,
-        isPending: false,
-        isError: true,
-        error: new Error("NOT_FOUND"),
-        refetch: vi.fn(),
-      });
-
-      render(
-        <Suspense fallback={null}>
-          <TripDetailContent tripId="trip-123" />
-        </Suspense>,
-      );
-
-      await waitFor(() => {
-        expect(screen.getByText("Trip not found")).toBeDefined();
-      });
-
-      expect(screen.queryByTestId("trip-notification-bell")).toBeNull();
-    });
-
-    it("does not render TripNotificationBell in loading state", () => {
-      mockUseTripDetail.mockReturnValue({
-        data: undefined,
-        isPending: true,
-        isError: false,
-        error: null,
-        refetch: vi.fn(),
-      });
-
-      render(
-        <Suspense fallback={null}>
-          <TripDetailContent tripId="trip-123" />
-        </Suspense>,
-      );
-
-      expect(screen.queryByTestId("trip-notification-bell")).toBeNull();
-    });
-  });
 });
