+++ apps/web/src/components/notifications/index.ts
@@ -1,7 +1,4 @@
 export { NotificationBell } from "./notification-bell";
-export { NotificationDropdown } from "./notification-dropdown";
 export { NotificationItem } from "./notification-item";
 export { NotificationPreferences } from "./notification-preferences";
-export { TripNotificationBell } from "./trip-notification-bell";
-export { TripNotificationDialog } from "./trip-notification-dialog";
 export { TripSettingsButton } from "./trip-settings-button";
